fixes delayed message parsing
`receive_message` was awaiting before parsing data alrady received,
causing it to always return messages with a lag if anything was
received immediately after the handshake.

Instead, it now parses as many messages as possible from
`self._msg_data` and only awaits for new data if no messages are found.